+ improved version

+++ selection_tools/build_selection.sh
@@ -6,7 +6,7 @@ WIKI=$LANGUAGE"wiki"
 
 if [ "$LANGUAGE" = '' ];
 then
-    echo "usage: WP1.sh <lang> (en for 'en' for example)"
+    echo "usage: build_selection.sh <lang> (en for 'en' for example)"
     exit
 fi
 
@@ -35,42 +35,69 @@ then
     LAST_VERSION=`cat ./tmp/$WIKI/date`
     if [ ! $LAST_VERSION = $CURRENT_VERSION ]
     then
-	rm ./tmp/$WIKI/date
-	rm ./tmp/$WIKI/source/* >& /dev/null
+	rm -f ./tmp/$WIKI/date >& /dev/null
+	rm -f ./tmp/$WIKI/source/* >& /dev/null
+	rm -f ./tmp/$WIKI/target/* >& /dev/null
+	rm -f ./tmp/$WIKI/target/.* >& /dev/null
     fi
 fi
 
 echo $CURRENT_VERSION > ./tmp/$WIKI/date
 
-## GET SQL DUMPS FROM download.wikimedia.org
-wget --continue -O ./tmp/$WIKI/source/$WIKI-latest-page.sql.gz http://download.wikimedia.org/$WIKI/latest/$WIKI-latest-page.sql.gz
-wget --continue -O ./tmp/$WIKI/source/$WIKI-latest-pagelinks.sql.gz http://download.wikimedia.org/$WIKI/latest/$WIKI-latest-pagelinks.sql.gz
-wget --continue -O ./tmp/$WIKI/source/$WIKI-latest-langlinks.sql.gz http://download.wikimedia.org/$WIKI/latest/$WIKI-latest-langlinks.sql.gz
-wget --continue -O ./tmp/$WIKI/source/$WIKI-latest-redirect.sql.gz http://download.wikimedia.org/$WIKI/latest/$WIKI-latest-redirect.sql.gz
-wget --continue -O ./tmp/$WIKI/source/$WIKI-latest-categorylinks.sql.gz http://download.wikimedia.org/$WIKI/latest/$WIKI-latest-categorylinks.sql.gz
+## GET SQL DUMPS FROM dumps.wikimedia.org
+echo "Downloading $WIKI-latest-page.sql.gz..."
+wget --no-verbose --continue -O ./tmp/$WIKI/source/$WIKI-latest-page.sql.gz http://dumps.wikimedia.org/$WIKI/latest/$WIKI-latest-page.sql.gz
 
-rm ./tmp/$WIKI/target/* >& /dev/null
+echo "Downloading $WIKI-latest-pagelinks.sql.gz..."
+wget --no-verbose --continue -O ./tmp/$WIKI/source/$WIKI-latest-pagelinks.sql.gz http://dumps.wikimedia.org/$WIKI/latest/$WIKI-latest-pagelinks.sql.gz
+
+echo "Downloading $WIKI-latest-langlinks.sql.gz..."
+wget --no-verbose --continue -O ./tmp/$WIKI/source/$WIKI-latest-langlinks.sql.gz http://dumps.wikimedia.org/$WIKI/latest/$WIKI-latest-langlinks.sql.gz
+
+echo "Download $WIKI-latest-redirect.sql.gz..."
+wget --no-verbose --continue -O ./tmp/$WIKI/source/$WIKI-latest-redirect.sql.gz http://dumps.wikimedia.org/$WIKI/latest/$WIKI-latest-redirect.sql.gz
 
 ## BUILD PAGES INDEXES
-cat ./tmp/$WIKI/source/$WIKI-latest-page.sql.gz | gzip -d | tail -n +38 | ./bin/pages_parser | egrep "^[0-9]+ 0 " | sort -n -t " " -k 1,1 | gzip > ./tmp/$WIKI/target/main_pages_sort_by_ids.lst.gz
-
-## BUILD TALK INDEXES
-cat ./tmp/$WIKI/source/$WIKI-latest-page.sql.gz | gzip -d | tail -n +38 | ./bin/pages_parser | egrep "^[0-9]+ 1 " | sort -n -t " " -k 1,1 | gzip > ./tmp/$WIKI/target/talk_pages_sort_by_ids.lst.gz
-
-## BUILD CATEGORIES INDEXES
-cat ./tmp/$WIKI/source/$WIKI-latest-page.sql.gz | gzip -d | tail -n +38 | ./bin/pages_parser | egrep "^[0-9]+ 14 " | sort -n -t " " -k 1,1 | gzip > ./tmp/$WIKI/target/categories_sort_by_ids.lst.gz
+echo "Computing ./tmp/$WIKI/target/main_pages_sort_by_ids.lst.gz..."
+if [ ! -f ./tmp/$WIKI/target/.main_pages_sort_by_ids.lst.gz.finished ]
+then
+    cat ./tmp/$WIKI/source/$WIKI-latest-page.sql.gz | gzip -d | tail -n +38 | ./bin/pages_parser | egrep "^[0-9]+ 0 " | sort -n -t " " -k 1,1 | gzip > ./tmp/$WIKI/target/main_pages_sort_by_ids.lst.gz
+    touch ./tmp/$WIKI/target/.main_pages_sort_by_ids.lst.gz.finished
+else
+    echo "./tmp/$WIKI/target/main_pages_sort_by_ids.lst.gz already computed."
+fi
 
 ## BUILD PAGELINKS INDEXES
-cat ./tmp/$WIKI/source/$WIKI-latest-pagelinks.sql.gz| gzip -d | tail -n +28 | ./bin/pagelinks_parser | gzip > ./tmp/$WIKI/target/pagelinks.lst.gz
+echo "Computing ./tmp/$WIKI/target/pagelinks.lst.gz..."
+if [ ! -f ./tmp/$WIKI/target/.pagelinks.lst.gz.finished ]
+then
+    cat ./tmp/$WIKI/source/$WIKI-latest-pagelinks.sql.gz| gzip -d | tail -n +28 | ./bin/pagelinks_parser | gzip > ./tmp/$WIKI/target/pagelinks.lst.gz
+    touch ./tmp/$WIKI/target/.pagelinks.lst.gz.finished
+else
+    echo "./tmp/$WIKI/target/pagelinks.lst.gz already computed."
+fi
 
 ## BUILD LANGLINKS INDEXES
-cat ./tmp/$WIKI/source/$WIKI-latest-langlinks.sql.gz | gzip -d | tail -n +28 | ./bin/langlinks_parser | sort -n -t " " -k 1,1 | gzip > ./tmp/$WIKI/target/langlinks_sort_by_ids.lst.gz
+echo "Computing ./tmp/$WIKI/target/langlinks_sort_by_ids.lst.gz..."
+if [ ! -f ./tmp/$WIKI/target/.langlinks_sort_by_ids.lst.gz.finished ]
+then
+    cat ./tmp/$WIKI/source/$WIKI-latest-langlinks.sql.gz | gzip -d | tail -n +28 | ./bin/langlinks_parser | sort -n -t " " -k 1,1 | gzip > ./tmp/$WIKI/target/langlinks_sort_by_ids.lst.gz
+    touch "./tmp/$WIKI/target/.langlinks_sort_by_ids.lst.gz.finished"
+else
+    echo "./tmp/$WIKI/target/langlinks_sort_by_ids.lst.gz already computed."
+fi
 
 ## BUILD REDIRECT INDEXES
-cat ./tmp/$WIKI/source/$WIKI-latest-redirect.sql.gz | gzip -d | tail -n +28 | ./bin/redirects_parser | sort -n -t " " -k 1,1 | gzip > ./tmp/$WIKI/target/redirects_sort_by_ids.lst.gz
+echo "Computing ./tmp/$WIKI/target/redirects_sort_by_ids.lst.gz..."
+if [ ! -f ./tmp/$WIKI/target/.redirects_sort_by_ids.lst.gz.finished ]
+then
+    cat ./tmp/$WIKI/source/$WIKI-latest-redirect.sql.gz | gzip -d | tail -n +28 | ./bin/redirects_parser | sort -n -t " " -k 1,1 | gzip > ./tmp/$WIKI/target/redirects_sort_by_ids.lst.gz
+    touch "./tmp/$WIKI/target/.redirects_sort_by_ids.lst.gz.finished"
+else
+    echo "./tmp/$WIKI/target/redirects_sort_by_ids.lst.gz already computed."
+fi
 
-## BUILD CATEGORYLINKS INDEXES
-cat ./tmp/$WIKI/source/$WIKI-latest-categorylinks.sql.gz | gzip -d | tail -n +28 | ./bin/categorylinks_parser | sort -n -t " " -k 1,1 | gzip > ./tmp/$WIKI/target/categorylinks_sort_by_ids.lst.gz
+exit
 
 ## BUILD CHARTS INDEXES
 ./bin/filter_charts.pl --chartsDirectory=./tmp/charts/ --language=$LANGUAGE | gzip > ./tmp/$WIKI/target/charts.lst.gz
